Remove thread safety
I felt it was best to leave thread safety up to the user of this package because in many cases it would be taking up unnecessary computing power. I'm using this package in one of my personal projects, and I found that the extra layer of thread safety wasn't really doing much to help.